@@ -2,18 +2,16 @@ import 'package:flutter/material.dart';
 import 'package:responsive_scaffold/responsive_scaffold.dart';
 
 class ListExample extends StatefulWidget {
-  const ListExample({
-    Key key,
-  });
+  const ListExample({Key? key}) : super(key: key);
 
   @override
   _ListExampleState createState() => _ListExampleState();
 }
 
 class _ListExampleState extends State<ListExample> {
-  var _scaffoldKey = new GlobalKey<ScaffoldState>();
+  final _scaffoldKey = GlobalKey<ScaffoldState>();
 
-  List<String> _items;
+  late List<String> _items;
 
   @override
   void initState() {
@@ -28,8 +26,7 @@ class _ListExampleState extends State<ListExample> {
       detailBuilder: (BuildContext context, int index, bool tablet) {
         final i = _items[index];
         return DetailsScreen(
-          body: new ExampleDetailsScreen(
-            items: _items,
+          body: ExampleDetailsScreen(
             row: i,
             tablet: tablet,
             onDelete: () {
@@ -46,14 +43,14 @@ class _ListExampleState extends State<ListExample> {
           ),
         );
       },
-      nullItems: Center(child: CircularProgressIndicator()),
-      emptyItems: Center(child: Text("No Items Found")),
-      slivers: <Widget>[
+      nullItems: const Center(child: CircularProgressIndicator()),
+      emptyItems: const Center(child: Text("No Items Found")),
+      slivers: const <Widget>[
         SliverAppBar(
           title: Text("App Bar"),
         ),
       ],
-      itemCount: _items?.length ?? 0,
+      itemCount: _items.length,
       itemBuilder: (BuildContext context, int index) {
         final i = _items[index];
         return ListTile(
@@ -62,20 +59,20 @@ class _ListExampleState extends State<ListExample> {
       },
       bottomNavigationBar: BottomAppBar(
         elevation: 0.0,
-        child: Container(
-          child: IconButton(
-            icon: Icon(Icons.share),
-            onPressed: () {},
-          ),
+        child: IconButton(
+          icon: const Icon(Icons.share),
+          onPressed: () {},
         ),
       ),
       floatingActionButton: FloatingActionButton(
-        child: Icon(Icons.add),
         onPressed: () {
-          _scaffoldKey.currentState.showSnackBar(SnackBar(
-            content: Text("Snackbar!"),
-          ));
+          ScaffoldMessenger.of(context).showSnackBar(
+            const SnackBar(
+              content: Text("Snackbar!"),
+            ),
+          );
         },
+        child: const Icon(Icons.add),
       ),
     );
   }
@@ -83,16 +80,14 @@ class _ListExampleState extends State<ListExample> {
 
 class ExampleDetailsScreen extends StatelessWidget {
   const ExampleDetailsScreen({
-    Key key,
-    @required List<String> items,
-    @required this.row,
-    @required this.tablet,
-    @required this.onDelete,
-    @required this.onChanged,
-  })  : _items = items,
+    Key? key,
+    required this.row,
+    required this.tablet,
+    required this.onDelete,
+    required this.onChanged,
+  })   : 
         super(key: key);
 
-  final List<String> _items;
   final String row;
   final bool tablet;
   final VoidCallback onDelete;
@@ -104,33 +99,29 @@ class ExampleDetailsScreen extends StatelessWidget {
       appBar: AppBar(
         elevation: 0.0,
         automaticallyImplyLeading: !tablet,
-        title: Text("Details"),
+        title: const Text("Details"),
         actions: [
           IconButton(
-            icon: Icon(Icons.share),
+            icon: const Icon(Icons.share),
             onPressed: () {
-              onChanged(row + " " + DateTime.now().toString());
+              onChanged("$row ${DateTime.now().toString()}");
             },
           ),
           IconButton(
-            icon: Icon(Icons.delete),
+            icon: const Icon(Icons.delete),
             onPressed: onDelete,
           ),
         ],
       ),
       bottomNavigationBar: BottomAppBar(
         elevation: 0.0,
-        child: Container(
-          child: IconButton(
-            icon: Icon(Icons.share),
-            onPressed: () {},
-          ),
+        child: IconButton(
+          icon: const Icon(Icons.share),
+          onPressed: () {},
         ),
       ),
-      body: Container(
-        child: Center(
-          child: Text("Item: $row"),
-        ),
+      body: Center(
+        child: Text("Item: $row"),
       ),
     );
   }
