@@ -4,9 +4,8 @@ import '../three_column_navigation.dart';
 
 class DetailsView extends StatelessWidget {
   const DetailsView({
-    Key key,
-    @required int listIndex,
-    @required DetailsWidget details,
+    Key? key,
+    required DetailsWidget details,
     this.previous,
     this.next,
     this.isLast = false,
@@ -14,17 +13,16 @@ class DetailsView extends StatelessWidget {
     this.automaticallyImplyLeading = true,
     this.showDetailsArrows = false,
     this.scaffoldKey,
-  })  : _listIndex = listIndex,
+  })  : 
         _details = details,
         super(key: key);
 
-  final int _listIndex;
   final DetailsWidget _details;
-  final VoidCallback previous, next;
+  final VoidCallback? previous, next;
   final bool showDetailsArrows;
   final bool isLast, isFirst;
   final bool automaticallyImplyLeading;
-  final GlobalKey<ScaffoldState> scaffoldKey;
+  final GlobalKey<ScaffoldState>? scaffoldKey;
 
   @override
   Widget build(BuildContext context) {
@@ -36,31 +34,31 @@ class DetailsView extends StatelessWidget {
             ? _details.title
             : Row(
                 mainAxisAlignment: MainAxisAlignment.center,
-                crossAxisAlignment: CrossAxisAlignment.center,
+                // crossAxisAlignment: CrossAxisAlignment.center,
                 children: <Widget>[
                   Row(
                     mainAxisSize: MainAxisSize.min,
                     children: <Widget>[
                       IconButton(
-                        icon: Icon(Icons.arrow_upward),
+                        icon: const Icon(Icons.arrow_upward),
                         onPressed: isFirst ? null : previous,
                       ),
                       IconButton(
-                        icon: Icon(Icons.arrow_downward),
+                        icon: const Icon(Icons.arrow_downward),
                         onPressed: isLast ? null : next,
                       ),
                     ],
                   ),
-                  if (_details?.title != null)
+                  if (_details.title != null)
                     Expanded(
                       child: Center(child: _details.title),
                     ),
                 ],
               ),
-        actions: _details?.actions,
+        actions: _details.actions,
       ),
       body: _details.child,
-      bottomNavigationBar: _details?.bottomAppBar,
+      bottomNavigationBar: _details.bottomAppBar,
     );
   }
 }
